fix(groupBy): return Partial Record
The grouping classifier function may have never returned one of the
possible values, so there is not necessarily an array for each group.

Note: The test is sort of copied from DefinitelyTyped.
I noticed you seem to use `tsd expect*` functions in the other test,
but wanted to avoid putting a huge `expectType` of sorts.
Testing this way seems like a pretty good idea here, because it's
user-focused ('this expression with the groupBy result is allowed, this
one is not').
Feel free to amend the test if you don't like that part.